feat: 브랜치 병합 과정에서의 충돌 해결
