+++ cart_service/src/main/java/on/ssgdeal/cart_service/exception/CartException.java
@@ -9,18 +9,21 @@ public class CartException extends CustomException {
     }
 
     public static class CartProductNotFoundException extends CustomException {
+
         public CartProductNotFoundException() {
             super(CartExceptionCode.CART_PRODUCT_NOT_FOUND);
         }
     }
 
     public static class NotEnoughStockException extends CartException {
+
         public NotEnoughStockException() {
             super(CartExceptionCode.NOT_ENOUGH_STOCK);
         }
     }
 
     public static class MustBePositiveQuantityException extends CartException {
+
         public MustBePositiveQuantityException() {
             super(CartExceptionCode.MUST_BE_POSITIVE_QUANTITY);
         }
